backend/requirements.txt
fastapi
uvicorn
transformers
torch
openai-whisper
pandas
python-multipart
